fix get cuisines

+++ client_end_user/shared/src/commonMain/kotlin/domain/entity/PaginationItems.kt
@@ -3,5 +3,5 @@ package domain.entity
 data class PaginationItems<T>(
     val items: List<T>,
     val page: Int,
-    val total: Int
+    val total: Long
 )
